Refactor DataCite Metadata adapter for WorkVerisons and Collections
Previously we had a class `DataCite::Metadata` that took a WorkVersion and turned it into the metadata schema that DataCite requires to publish a DOI.

With the advent of minting DOIs for Collections, I needed to write a corresponding metadata adapter to translate a Collection's metdata into datacite format.
