+++ lib/data_cite.rb
@@ -2,5 +2,7 @@
 
 module DataCite
   require 'data_cite/client'
-  require 'data_cite/metadata'
+  require 'data_cite/metadata/base'
+  require 'data_cite/metadata/work_version'
+  require 'data_cite/metadata/collection'
 end
