+++ lib/data_cite/metadata/collection.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module DataCite
+  module Metadata
+    class Collection < Base
+      private
+
+        def resource_type
+          'Collection'
+        end
+    end
+  end
+end
